requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.0
strands-agents
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
requests==2.31.0
